provision for a 4th ntuple layout used by LBNE
git-svn-id: file:///home/mroda/Software/Genie/NewGlobal/trunk@3471 cc9776de-3512-45ca-aafc-e2d9ed43c22c

+++ src/FluxDrivers/GNuMINtuple/mkgclasses3.sh
@@ -2,21 +2,29 @@
 
 do_mkclass="yes"
 
-for fbase in g3numi g4numi flugg
+for fbase in g3numi g4numi flugg g4lbne
 do
   tname="xyzzy"
   if [ "$fbase" = "g3numi" ] ; then tname="h10" ; fi
   if [ "$fbase" = "flugg"  ] ; then tname="h10" ; fi
   if [ "$fbase" = "g4numi" ] ; then tname="nudata" ; fi
+  if [ "$fbase" = "g4lbne" ] ; then tname="nudata" ; fi
 
   if [ "yes" = "${do_mkclass}" ] ; then
     cat > make_generic.C <<EOF
+#include <string>
+#include <vector>
+#include <map>
 void make_generic()
 {
   TFile* g3file = TFile::Open("generic_${fbase}.root","READONLY");
   gROOT->ProcessLine("${tname}->MakeClass();");
 }
 EOF
+    if [ ! -f generic_${fbase}.root ]; then
+      echo "skip $fbase for lack of generic_${fbase}.root"
+      continue
+    fi
     root -b -q make_generic.C
     rm -f make_generic.C
   fi
